Logistic Regression in R
Logistic regression is a classification Algorithm in machine learning where you will predict the categories it is linear model. This code will help us predict the users in social media who would like to buy new SUV for a car company so that the car company can advertise and target those users in the social media.